+++ jubeatools/formats/jubeat_analyser/__init__.py
@@ -1,12 +1,17 @@
 """
-This module contains code for the different formats read by jubeat analyser
+This module contains code for the different formats read by "jubeat analyser".
 
-"memo" is a vague term that refers to several legacy formats.
-They were originally derived from the (somewhat) human-readable format choosen
-by websites gathering official jubeat charts in text form as a memory aid.
+"jubeat analyser" is a Windows program that can play back chart files and
+export them to video files, it can also be used as a jubeat simulator.
 
-The machine-readable variants are partially documented (in japanese)
-on these pages :
+"memo" is a vague term that refers to several slightly different formats.
+My understanding is that they were all originally derived from the (somewhat)
+human-readable format choosen by websites like jubeat memo or cosmos memo.
+These websites would provide text transcripts of official jubeat charts as
+training material for hardcore players.
+
+The machine-readable variants or these text formats are partially documented
+(in japanese) on these pages :
 - http://yosh52.web.fc2.com/jubeat/fumenformat.html
 - http://yosh52.web.fc2.com/jubeat/holdmarker.html
 """